+++ worlds/banjo_tooie/Names/locationName.py
@@ -1,0 +1,274 @@
+#Doubloons
+JRLDB1  = "Jolly Rogers: Town Center Pole 1 Doubloon"
+JRLDB2  = "Jolly Rogers: Town Center Pole 2 Doubloon"
+JRLDB3  = "Jolly Rogers: Town Center Pole 3 Doubloon"
+JRLDB4  = "Jolly Rogers: Town Center Pole 4 Doubloon"
+JRLDB5  = "Jolly Rogers: Town Center Pole 5 Doubloon"
+JRLDB6  = "Jolly Rogers: Town Center Pole 6 Doubloon"
+JRLDB7  = "Jolly Rogers: Silo 1 Doubloon"
+JRLDB8  = "Jolly Rogers: Silo 2 Doubloon"
+JRLDB9  = "Jolly Rogers: Silo 3 Doubloon"
+JRLDB10 = "Jolly Rogers: Silo 4 Doubloon"
+JRLDB11 = "Jolly Rogers: Toxic Pool 1 Doubloon"
+JRLDB12 = "Jolly Rogers: Toxic Pool 2 Doubloon"
+JRLDB13 = "Jolly Rogers: Toxic Pool 3 Doubloon"
+JRLDB14 = "Jolly Rogers: Toxic Pool 4 Doubloon"
+JRLDB15 = "Jolly Rogers: Mumbo Skull 1 Doubloon"
+JRLDB16 = "Jolly Rogers: Mumbo Skull 2 Doubloon"
+JRLDB17 = "Jolly Rogers: Mumbo Skull 3 Doubloon"
+JRLDB18 = "Jolly Rogers: Mumbo Skull 4 Doubloon"
+JRLDB19 = "Jolly Rogers: Underground 1 Doubloon"
+JRLDB20 = "Jolly Rogers: Underground 2 Doubloon"
+JRLDB21 = "Jolly Rogers: Underground 3 Doubloon"
+JRLDB22 = "Jolly Rogers: Alcove 1 Doubloon"
+JRLDB23 = "Jolly Rogers: Alcove 2 Doubloon"
+JRLDB24 = "Jolly Rogers: Alcove 3 Doubloon"
+JRLDB25 = "Jolly Rogers: Capt Blackeye 1 Doubloon"
+JRLDB26 = "Jolly Rogers: Capt Blackeye 2 Doubloon"
+JRLDB27 = "Jolly Rogers: Near Jinjo 1 Doubloon"
+JRLDB28 = "Jolly Rogers: Near Jinjo 2 Doubloon"
+JRLDB29 = "Jolly Rogers: Near Jinjo 3 Doubloon"
+JRLDB30 = "Jolly Rogers: Near Jinjo 4 Doubloon"
+
+#Jinjo
+JINJOMT1 = "Mayahem Temple: Jade Snake Grove Jinjo"
+JINJOMT2 = "Mayahem Temple: Stadium Jinjo"
+JINJOMT3 = "Mayahem Temple: Targitzan Temple Jinjo"
+JINJOMT4 = "Mayahem Temple: Water Pool Jinjo"
+JINJOMT5 = "Mayahem Temple: Bridge Jinjo"
+JINJOGM1 = "Glitter Gultch Mine: Water Storage Jinjo"
+JINJOGM2 = "Glitter Gultch Mine: Jail Jinjo"
+JINJOGM3 = "Glitter Gultch Mine: Toxic Gas Cave Jinjo"
+JINJOGM4 = "Glitter Gultch Mine: Boulder Jinjo"
+JINJOGM5 = "Glitter Gultch Mine: Mine Tracks Jinjo"
+JINJOWW1 = "Witchyworld: Big Top Jinjo"
+JINJOWW2 = "Witchyworld: Cave of Horrors Jinjo"
+JINJOWW3 = "Witchyworld: Van Door Jinjo"
+JINJOWW4 = "Witchyworld: Dodgem Dome Jinjo"
+JINJOWW5 = "Witchyworld: Cactus of Strength Jinjo"
+JINJOJR1 = "Jolly Rogers: Lagoon Alcove Jinjo"
+JINJOJR2 = "Jolly Rogers: Blubber Jinjo"
+JINJOJR3 = "Jolly Rogers: Big Fish Jinjo"
+JINJOJR4 = "Jolly Rogers: Seaweed Sanctum Jinjo"
+JINJOJR5 = "Jolly Rogers: Sunken Ship Jinjo"
+JINJOTL1 = "Terrydactyland: Talon Torp Jinjo"
+JINJOTL2 = "Terrydactyland: Cutscene Skip Jinjo"
+JINJOTL3 = "Terrydactyland: Beside Rocknut Jinjo"
+JINJOTL4 = "Terrydactyland: Big T. Rex Skip Jinjo"
+JINJOTL5 = "Terrydactyland: Stomping Plains Jinjo"
+JINJOGI1 = "Gruntys Industries: Floor 5 Jinjo"
+JINJOGI2 = "Gruntys Industries: Leg Spring Jinjo"
+JINJOGI3 = "Gruntys Industries: Waste Disposal Plant Jinjo"
+JINJOGI4 = "Gruntys Industries: Boiler Plant Jinjo"
+JINJOGI5 = "Gruntys Industries: Outside Jinjo"
+JINJOHP1 = "Hailfire Peaks: Lava Waterfall Jinjo"
+JINJOHP2 = "Hailfire Peaks: Boiling Hot Pool Jinjo"
+JINJOHP3 = "Hailfire Peaks: Windy Hole Jinjo"
+JINJOHP4 = "Hailfire Peaks: Icicle Grotto Jinjo"
+JINJOHP5 = "Hailfire Peaks: Mildred Ice Cube Jinjo"
+JINJOCC1 = "Cloud Cuckcooland: Trash Can Jinjo"
+JINJOCC2 = "Cloud Cuckcooland: Cheese Wedge Jinjo"
+JINJOCC3 = "Cloud Cuckcooland: Central Cavern Jinjo"
+JINJOCC4 = "Cloud Cuckcooland: Fake Mumbo Skull Jinjo"
+JINJOCC5 = "Cloud Cuckcooland: Wumba Jinjo"
+JINJOIH1 = "Wooded Hollow: Jinjo"
+JINJOIH2 = "Wasteland: Jinjo"
+JINJOIH3 = "Cliff Top: Jinjo"
+JINJOIH4 = "Plateau: Jinjo"
+JINJOIH5 = "Spiral Mountain: Jinjo"
+
+#Jiggies
+JIGGYMT1  = "MT: Targitzan Jiggy"
+JIGGYMT2  = "MT: Slightly Sacred Chamber Jiggy"
+JIGGYMT3  = "MT: Kickball Jiggy"
+JIGGYMT4  = "MT: Bovina Jiggy"
+JIGGYMT5  = "MT: Treasure Chamber Jiggy"
+JIGGYMT6  = "MT: Golden Goliath Jiggy"
+JIGGYMT7  = "MT: Prison Compound Quicksand Jiggy"
+JIGGYMT8  = "MT: Pillars Jiggy"
+JIGGYMT9  = "MT: Top of Temple Jiggy"
+JIGGYMT10 = "MT: Ssslumber Jiggy"
+JIGGYGM1  = "GGM: Old King Coal Jiggy"
+JIGGYGM2  = "GGM: Canary Mary Jiggy"
+JIGGYGM3  = "GGM: Generator Cavern Jiggy"
+JIGGYGM4  = "GGM: Waterfall Cavern Jiggy"
+JIGGYGM5  = "GGM: Ordnance Storage Jiggy"
+JIGGYGM6  = "GGM: Dilberta Jiggy"
+JIGGYGM7  = "GGM: Crushing Shed Jiggy"
+JIGGYGM8  = "GGM: Waterfall Jiggy"
+JIGGYGM9  = "GGM: Power Hut Basement Jiggy"
+JIGGYGM10 = "GGM: Flooded Caves Jiggy"
+JIGGYWW1  = "WW: Hoop Hurry Jiggy"
+JIGGYWW2  = "WW: Dodgem Dome Jiggy"
+JIGGYWW3  = "WW: Mr. Patch Jiggy"
+JIGGYWW4  = "WW: Saucer of Peril Jiggy"
+JIGGYWW5  = "WW: Balloon Burst Jiggy"
+JIGGYWW6  = "WW: Dive of Death Jiggy"
+JIGGYWW7  = "WW: Mrs. Boggy Jiggy"
+JIGGYWW8  = "WW: Star Spinner Jiggy"
+JIGGYWW9  = "WW: The Inferno Jiggy"
+JIGGYWW10 = "WW: Cactus of Strength Jiggy"
+JIGGYJR1  = "JRL: Mini-Sub Challenge Jiggy"
+JIGGYJR2  = "JRL: Tiptup Jiggy"
+JIGGYJR3  = "JRL: Chris P. Bacon Jiggy"
+JIGGYJR4  = "JRL: Pig Pool Jiggy"
+JIGGYJR5  = "JRL: Smuggler's Cavern Jiggy"
+JIGGYJR6  = "JRL: Merry Maggie Jiggy"
+JIGGYJR7  = "JRL: Woo Fak Fak Jiggy"
+JIGGYJR8  = "JRL: Seemee Jiggy"
+JIGGYJR9  = "JRL: Pawno Jiggy"
+JIGGYJR10 = "JRL: UFO Jiggy"
+JIGGYTD1  = "TDL: Under Terry's Nest Jiggy"
+JIGGYTD2  = "TDL: Dippy Jiggy"
+JIGGYTD3  = "TDL: Scrotty Jiggy"
+JIGGYTD4  = "TDL: Terry Jiggy"
+JIGGYTD5  = "TDL: Oogle Boogle Tribe Jiggy"
+JIGGYTD6  = "TDL: Chompas Belly Jiggy"
+JIGGYTD7  = "TDL: Terry's Kids Jiggy"
+JIGGYTD8  = "TDL: Stomping Plains Jiggy"
+JIGGYTD9  = "TDL: Rocknut Tribe Jiggy"
+JIGGYTD10 = "TDL: Code of the Dinosaurs Jiggy"
+JIGGYGI1  = "GI: Underwater Waste Disposal Plant Jiggy"
+JIGGYGI2  = "GI: Weldar Jiggy"
+JIGGYGI3  = "GI: Clinker's Cavern Jiggy"
+JIGGYGI4  = "GI: Skivvy Workers Jiggy"
+JIGGYGI5  = "GI: Floor 5 Jiggy"
+JIGGYGI6  = "GI: Quality Control Jiggy"
+JIGGYGI7  = "GI: Floor 1 Guarded Jiggy"
+JIGGYGI8  = "GI: Trash Compactor Jiggy"
+JIGGYGI9  = "GI: Twinkly Packing Jiggy"
+JIGGYGI10 = "GI: Waste Disposal Plant Box Jiggy"
+JIGGYHP1  = "HFP: Dragon Brothers Jiggy"
+JIGGYHP2  = "HFP: Inside the Volcano Jiggy"
+JIGGYHP3  = "HFP: Sabreman Jiggy"
+JIGGYHP4  = "HFP: Boggy Jiggy"
+JIGGYHP5  = "HFP: Icy Side Station Jiggy"
+JIGGYHP6  = "HFP: Oil Drill Jiggy"
+JIGGYHP7  = "HFP: Stomping Plains Jiggy"
+JIGGYHP8  = "HFP: Kickball Jiggy"
+JIGGYHP9  = "HFP: Aliens Jiggy"
+JIGGYHP10 = "HFP: Colosseum Split-Up Jiggy"
+JIGGYCC1  = "CCL: Mingy Jongo Jiggy"
+JIGGYCC2  = "CCL: Mr Fit Jiggy"
+JIGGYCC3  = "CCL: Pot O' Gold Jiggy"
+JIGGYCC4  = "CCL: Canary Mary Jiggy"
+JIGGYCC5  = "CCL: Zubbas Jiggy"
+JIGGYCC6  = "CCL: Jiggium Plant Jiggy"
+JIGGYCC7  = "CCL: Cheese Wedge Jiggy"
+JIGGYCC8  = "CCL: Trash Can Jiggy"
+JIGGYCC9  = "CCL: Superstash Jiggy"
+JIGGYCC10 = "CCL: Jelly Castle Jiggy"
+JIGGYIH1  = "Jinjo Village: White Jinjo Family Jiggy"
+JIGGYIH2  = "Jinjo Village: Orange Jinjo Family Jiggy"
+JIGGYIH3  = "Jinjo Village: Yellow Jinjo Family Jiggy"
+JIGGYIH4  = "Jinjo Village: Brown Jinjo Family Jiggy"
+JIGGYIH5  = "Jinjo Village: Green Jinjo Family Jiggy"
+JIGGYIH6  = "Jinjo Village: Red Jinjo Family Jiggy"
+JIGGYIH7  = "Jinjo Village: Blue Jinjo Family Jiggy"
+JIGGYIH8  = "Jinjo Village: Purple Jinjo Family Jiggy"
+JIGGYIH9  = "Jinjo Village: Black Jinjo Family Jiggy"
+JIGGYIH10 = "Jinjo Village: King Jingaling Jiggy"
+
+#Glowbos
+GLOWBOMT1 = "MT: Mumbo's Glowbo"
+GLOWBOMT2 = "MT: Jade Snake Grove Glowbo"
+GLOWBOGM1 = "GGM: Near Entrance Glowbo"
+GLOWBOGM2 = "GGM: Mine Entrance 2 Glowbo"
+GLOWBOWW1 = "WW: The Inferno Glowbo"
+GLOWBOWW2 = "WW: Wumba's Glowbo"
+GLOWBOJR1 = "JRL: Pawno's Emporium Glowbo"
+GLOWBOJR2 = "JRL: Under Wumba's Wigwam Glowbo"
+GLOWBOTL1 = "TDL: Unga Bunga Cave Entrance Glowbo"
+GLOWBOTL2 = "TDL: Behind Mumbo's Skull Glowbo"
+GLOWBOGI1 = "GI: Floor 2 Glowbo"
+GLOWBOGI2 = "GI: Floor 3 Glowbo"
+GLOWBOHP1 = "HFP: Lava Side Glowbo"
+GLOWBOHP2 = "HFP: Icy Side Glowbo"
+GLOWBOCC1 = "CCL: Green Pool Glowbo"
+GLOWBOCC2 = "CCL: Central Cavern Glowbo"
+GLOWBOIH1 = "IoH: Cliff Top Glowbo"
+GLOWBOMEG = "Mega Glowbo"
+
+#Honeycombs
+HONEYCMT1  = "MT: Entrance Honeycomb"
+HONEYCMT2  = "MT: Bovina Honeycomb"
+HONEYCMT3  = "MT: Treasure Chamber Honeycomb"
+HONEYCGM1  = "GGM: Toxic Gas Cave Honeycomb"
+HONEYCGM2  = "GGM: Boulder Honeycomb"
+HONEYCGM3  = "GGM: Train Station Honeycomb"
+HONEYCWW1  = "WW: Space Zone Honeycomb"
+HONEYCWW2  = "WW: Mumbo Skull Honeycomb"
+HONEYCWW3  = "WW: Crazy Castle Honeycomb"
+HONEYCJR1  = "JRL: Seemee Honeycomb"
+HONEYCJR2  = "JRL: Atlantis Honeycomb"
+HONEYCJR3  = "JRL: Waste Pipe Honeycomb"
+HONEYCTL1  = "TDL: Lakeside Honeycomb"
+HONEYCTL2  = "TDL: Styracosaurus Cave Honeycomb"
+HONEYCTL3  = "TDL: River Passage Honeycomb"
+HONEYCGI1  = "GI: Floor 3 Honeycomb"
+HONEYCGI2  = "GI: Train Station Honeycomb"
+HONEYCGI3  = "GI: Chimney Honeycomb"
+HONEYCHP1  = "HFP: Inside the Volcano Honeycomb"
+HONEYCHP2  = "HFP: Train Station Honeycomb"
+HONEYCHP3  = "HFP: Lava Side Honeycomb"
+HONEYCCC1  = "CCL: Dirt Patch Honeycomb"
+HONEYCCC2  = "CCL: Trash Can Honeycomb"
+HONEYCCC3  = "CCL: Pot O' Gold Honeycomb"
+HONEYCIH1  = "IoH: Plateau Honeycomb"
+
+#Cheato Pages
+CHEATOMT1 = "MT: Snake Head Cheato Page"
+CHEATOMT2 = "MT: Prison Compound Cheato Page"
+CHEATOMT3 = "MT: Jade Snake Grove Cheato Page"
+CHEATOGM1 = "GGM: Canary Mary Cheato Page"
+CHEATOGM2 = "GGM: Entrance Cheato Page"
+CHEATOGM3 = "GGM: Water Storage Cheato Page"
+CHEATOWW1 = "WW: The Haunted Cavern Cheato Page"
+CHEATOWW2 = "WW: The Inferno Cheato Page"
+CHEATOWW3 = "WW: Saucer of Peril Cheato Page"
+CHEATOJR1 = "JRL: Pawno's Cheato Page"
+CHEATOJR2 = "JRL: Seemee Cheato Page"
+CHEATOJR3 = "JRL: Ancient Swimming Baths Cheato Page"
+CHEATOTL1 = "TDL: Dippy's Pool Cheato Page"
+CHEATOTL2 = "TDL: Inside the Mountain Cheato Page"
+CHEATOTL3 = "TDL: Boulder Cheato Page"
+CHEATOGI1 = "GI: Loggo Cheato Page"
+CHEATOGI2 = "GI: Floor 2 Window Cheato Page"
+CHEATOGI3 = "GI: Repair Depot Cheato Page"
+CHEATOHP1 = "HFP: Colosseum Cheato Page"
+CHEATOHP2 = "HFP: Icicle Grotto Cheato Page"
+CHEATOHP3 = "HFP: Icy Side Pillar Cheato Page"
+CHEATOCC1 = "CCL: Canary Mary Cheato Page"
+CHEATOCC2 = "CCL: Pot O' Gold Cheato Page"
+CHEATOCC3 = "CCL: Zubbas Cheato Page"
+CHEATOSM1 = "Spiral Mountain: Cheato Page"
+
+#Abilities
+GGRAB       = "MT: Grip Grab Silo"
+BBLASTER    = "MT: Breegull Blaster Silo"
+EGGAIM      = "MT: Egg Aim Silo"
+BDRILL      = "GGM: Bill Drill Silo"
+BBAYONET    = "GGM: Beak Bayonet Silo"
+AIREAIM     = "WW: Airborne Egg Aiming Silo"
+SPLITUP     = "WW: Split Up Silo"
+WWHACK      = "JRL: Wing Whack Silo"
+TTORP       = "JRL: Talon Torpedo Silo"
+AUQAIM      = "JRL: Sub-Aqua Egg Aiming Silo"
+SHPACK      = "HFP: Shack Pack Silo"
+GLIDE       = "HFP: Glide Silo"
+SNPACK      = "GI: Snooze Pack Silo"
+LSPRING     = "GI: Leg Spring Silo"
+CLAWBTS     = "GI: Claw Clamber Boots Silo"
+SPRINGB     = "TDL: Springy Step Shoes Silo"
+TAXPACK     = "TDL: Taxi Pack Silo"
+HATCH       = "TDL: Hatch Silo"
+PACKWH      = "WW: Pack Whack Silo"
+SAPACK      = "CCL: Sack Pack Silo"
+FEGGS       = "IoH: Fire Eggs Silo"
+GEGGS       = "IoH: Grenade Eggs Silo"
+CEGGS       = "IoH: Clockwork Kazooie Eggs Silo"
+IEGGS       = "IoH: Ice Eggs Silo"
+FSWIM       = "Fast Swimming"
+
+#Bosses
+HAG1 = "Hag 1 Defeated"
