+++ worlds/banjo_tooie/__init__.py
@@ -3,15 +3,12 @@ from multiprocessing import Process
 import settings
 import typing
 from jinja2 import Environment, FileSystemLoader
-# from .Game import game_name, filler_item_name
 from .Items import BanjoTooieItem, all_item_table, all_group_table
 from .Locations import BanjoTooieLocation, all_location_table
 from .Regions import BANJOTOOIEREGIONS, create_regions, connect_regions
 from .Options import BanjoTooieOptions
 from .Rules import BanjoTooieRules
 from .Names import itemName
-
-
 
 #from Utils import get_options
 from BaseClasses import ItemClassification, Tutorial, Item, Region, MultiWorld
@@ -25,11 +22,10 @@ def run_client():
     p = Process(target=main)
     p.start()
 
-
 components.append(Component("Banjo-Tooie Client", func=run_client, component_type=Type.CLIENT))
 
 class BanjoTooieWeb(WebWorld):
-    setup = Tutorial("Setup Banjo Tooie",
+    setup = Tutorial("Setup Banjo-Tooie",
         """Class to build website tutorial pages from a .md file in the world's /docs folder. Order is as follows.
         Name of the tutorial as it will appear on the site. Concise description covering what the guide will entail.
         Language the guide is written in. Name of the file ex 'setup_en.md'. Name of the link on the site; game name is
@@ -47,7 +43,7 @@ class BanjoTooieWorld(World):
     the game features three-dimensional worlds consisting of various platforming challenges and puzzles, with a notable
     increased focus on puzzle-solving over the worlds of Banjo-Kazooie.
     """
-    game: str = "Banjo Tooie"
+    game: str = "Banjo-Tooie"
     web = BanjoTooieWeb()
     options_dataclass =  BanjoTooieOptions
     options: BanjoTooieOptions
