--- worlds/banjotooie/Names/itemName.py
@@ -1,155 +1,0 @@
-PAGES = "Cheato Page"
-HONEY = "Empty Honeycomb"
-WJINJO = "white Jinjo"
-OJINJO = "Orange Jinjo"
-YJINJO = "Yellow Jinjo"
-BRJINJO = "Brown Jinjo"
-GJINJO = "Green Jinjo"
-RJINJO = "Red Jinjo"
-BLJINJO = "Blue Jinjo"
-PJINJO = "Purple Jinjo"
-BKJINJO = "Black Jinjo"
-DOUBLOON = "Doubloon"
-GLOWBO = "Glowbo"
-JIGGY = "Jiggy"
-
-#Jiggies
-JIGGYMT1  = "Jiggy 1"
-JIGGYMT2  = "Jiggy 2"
-JIGGYMT3  = "Jiggy 3"
-JIGGYMT4  = "Jiggy 4"
-JIGGYMT5  = "Jiggy 5"
-JIGGYMT6  = "Jiggy 6"
-JIGGYMT7  = "Jiggy 7"
-JIGGYMT8  = "Jiggy 8"
-JIGGYMT9  = "Jiggy 9"
-JIGGYMT10 = "Jiggy 10"
-JIGGYGM1  = "Jiggy 11"
-JIGGYGM2  = "Jiggy 12"
-JIGGYGM3  = "Jiggy 13"
-JIGGYGM4  = "Jiggy 14"
-JIGGYGM5  = "Jiggy 15"
-JIGGYGM6  = "Jiggy 16"
-JIGGYGM7  = "Jiggy 17"
-JIGGYGM8  = "Jiggy 18"
-JIGGYGM9  = "Jiggy 19"
-JIGGYGM10 = "Jiggy 20"
-JIGGYWW1  = "Jiggy 21"
-JIGGYWW2  = "Jiggy 22"
-JIGGYWW3  = "Jiggy 23"
-JIGGYWW4  = "Jiggy 24"
-JIGGYWW5  = "Jiggy 25"
-JIGGYWW6  = "Jiggy 26"
-JIGGYWW7  = "Jiggy 27"
-JIGGYWW8  = "Jiggy 28"
-JIGGYWW9  = "Jiggy 29"
-JIGGYWW10 = "Jiggy 30"
-JIGGYJR1  = "Jiggy 31"
-JIGGYJR2  = "Jiggy 32"
-JIGGYJR3  = "Jiggy 33"
-JIGGYJR4  = "Jiggy 34"
-JIGGYJR5  = "Jiggy 35"
-JIGGYJR6  = "Jiggy 36"
-JIGGYJR7  = "Jiggy 37"
-JIGGYJR8  = "Jiggy 38"
-JIGGYJR9  = "Jiggy 39"
-JIGGYJR10 = "Jiggy 40"
-JIGGYTD1  = "Jiggy 41"
-JIGGYTD2  = "Jiggy 42"
-JIGGYTD3  = "Jiggy 43"
-JIGGYTD4  = "Jiggy 44"
-JIGGYTD5  = "Jiggy 45"
-JIGGYTD6  = "Jiggy 46"
-JIGGYTD7  = "Jiggy 47"
-JIGGYTD8  = "Jiggy 48"
-JIGGYTD9  = "Jiggy 49"
-JIGGYTD10 = "Jiggy 50"
-JIGGYGI1  = "Jiggy 51"
-JIGGYGI2  = "Jiggy 52"
-JIGGYGI3  = "Jiggy 53"
-JIGGYGI4  = "Jiggy 54"
-JIGGYGI5  = "Jiggy 55"
-JIGGYGI6  = "Jiggy 56"
-JIGGYGI7  = "Jiggy 57"
-JIGGYGI8  = "Jiggy 58"
-JIGGYGI9  = "Jiggy 59"
-JIGGYGI10 = "Jiggy 60"
-JIGGYHP1  = "Jiggy 61"
-JIGGYHP2  = "Jiggy 62"
-JIGGYHP3  = "Jiggy 63"
-JIGGYHP4  = "Jiggy 64"
-JIGGYHP5  = "Jiggy 65"
-JIGGYHP6  = "Jiggy 66"
-JIGGYHP7  = "Jiggy 67"
-JIGGYHP8  = "Jiggy 68"
-JIGGYHP9  = "Jiggy 69"
-JIGGYHP10 = "Jiggy 70"
-JIGGYCC1  = "Jiggy 71"
-JIGGYCC2  = "Jiggy 72"
-JIGGYCC3  = "Jiggy 73"
-JIGGYCC4  = "Jiggy 74"
-JIGGYCC5  = "Jiggy 75"
-JIGGYCC6  = "Jiggy 76"
-JIGGYCC7  = "Jiggy 77"
-JIGGYCC8  = "Jiggy 78"
-JIGGYCC9  = "Jiggy 79"
-JIGGYCC10 = "Jiggy 80"
-JIGGYIH1  = "Jiggy 81"
-JIGGYIH2  = "Jiggy 82"
-JIGGYIH3  = "Jiggy 83"
-JIGGYIH4  = "Jiggy 84"
-JIGGYIH5  = "Jiggy 85"
-JIGGYIH6  = "Jiggy 86"
-JIGGYIH7  = "Jiggy 87"
-JIGGYIH8  = "Jiggy 88"
-JIGGYIH9  = "Jiggy 89"
-JIGGYIH10 = "Jiggy 90"
-
-GGRAB       = "Grip Grab"
-BBLASTER    = "Breegull Blaster"
-EGGAIM      = "Egg Aim"
-BDRILL      = "Bill Drill"
-BBAYONET    = "Beak Bayonet"
-AIREAIM     = "Airborne Egg Aiming" #TODO: Double Check
-SPLITUP     = "Split Up" #TODO: Double Check
-WWHACK      = "Wing Whack" #TODO: Double Check
-TTORP       = "Talon Torpedo" #TODO: Double Check
-AUQAIM      = "Sub-Aqua Egg Aiming" #TODO: Double Check
-SHPACK      = "Shack Pack"
-GLIDE       = "Glide"
-SNPACK      = "Snooze Pack"
-LSPRING     = "Leg Spring"
-CLAWBTS     = "Claw Clamber Boots"
-SPRINGB     = "Springy Step Shoes"
-TAXPACK     = "Taxi Pack"
-HATCH       = "Hatch"
-PACKWH      = "Pack Whack"
-SAPACK      = "Sack Pack"
-FEGGS       = "Fire Eggs"
-GEGGS       = "Grenade Eggs"
-CEGGS       = "Clockwork Kazooie Eggs"
-IEGGS       = "Ice Eggs"
-FSWIM       = "Fast Swimming"
-
-# Mumbo Magics
-MUMBOMT = "Mumbo: Golden Goliath"
-MUMBOGM = "Mumbo: Levitate"
-MUMBOWW = "Mumbo: Power"
-MUMBOJR = "Mumbo: Oxygenate"
-MUMBOTD = "Mumbo: Enlarge"
-MUMBOGI = "Mumbo: EMP"
-MUMBOHP = "Mumbo: Life Force"
-MUMBOCC = "Mumbo: Rain Dance"
-MUMBOIH = "Mumbo: Heal"
-
-HUMBAMT = "Humba: Stony"
-HUMBAGM = "Humba: Detonator"
-HUMBAWW = "Humba: Money Van"
-HUMBAJR = "Humba: Sub"
-HUMBATD = "Humba: T-Rex"
-HUMBAGI = "Humba: Washing Machine"
-HUMBAHP = "Humba: Snowball"
-HUMBACC = "Humba: Bee"
-HUMBAIH = "Humba: Dragon"
-
-VICTORY = "Kick Around"
