+++ yaml-template/template.yaml
@@ -1,4 +1,4 @@
-Banjo Tooie:
+Banjo-Tooie:
   death_link: 'false'
   multiworld_moves: 'true'
   multiworld_jinjos: 'false'
